revert 2 modules
